refactor: TokenProvider 추상화
