+++ src/test/java/roomescape/infrastructure/TokenProviderTest.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.infrastructure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -6,16 +6,17 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import roomescape.auth.TokenProvider;
 import roomescape.auth.exception.AuthorizationException;
 
-class TokenProviderTest {
+class JwtProviderTest {
     private static final String SECRET = "secretsecretsecretsecretsecretsecretsecretsecret";
 
-    private TokenProvider tokenProvider = new TokenProvider(SECRET, 10000L);
+    private TokenProvider tokenProvider = new JwtProvider(SECRET, 10000L);
 
     @Test
     void 토큰이_만료되면_예외가_발생한다() {
-        tokenProvider = new TokenProvider(SECRET, 0L);
+        tokenProvider = new JwtProvider(SECRET, 0L);
         String token = tokenProvider.createToken("1");
 
         assertThatThrownBy(() -> tokenProvider.extractSubject(token))
@@ -34,7 +35,7 @@ class TokenProviderTest {
     @Test
     void 올바르지_않은_토큰이면_예외가_발생한다() {
         String invalidSecret = "invalidinvalidinvalidinvalidinvalidinvalidinvalidinvalid";
-        TokenProvider invalidTokenProvider = new TokenProvider(invalidSecret, 10000L);
+        TokenProvider invalidTokenProvider = new JwtProvider(invalidSecret, 10000L);
         String invalidToken = invalidTokenProvider.createToken("1");
 
         assertThatThrownBy(() -> tokenProvider.extractSubject(invalidToken))
